remove disk space check instructions
